refactor(config): extract rebasing helpers
